Move GlobalizeResource to the Controls.Infrastructure namespace
