+++ src/DotVVM.Framework/Controls/Infrastructure/GlobalizeResource.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using DotVVM.Framework.Hosting;
 
-namespace DotVVM.Framework.Controls
+namespace DotVVM.Framework.Controls.Infrastructure
 {
     public class GlobalizeResource : DotvvmControl
     {
